refactor(bower): Migrate to the dependency graph API
