+++ plugins/package-managers/bower/src/main/kotlin/Bower.kt
@@ -24,21 +24,11 @@ import java.util.LinkedList
 
 import org.ossreviewtoolkit.analyzer.AbstractPackageManagerFactory
 import org.ossreviewtoolkit.analyzer.PackageManager
-import org.ossreviewtoolkit.analyzer.PackageManager.Companion.processProjectVcs
-import org.ossreviewtoolkit.analyzer.parseAuthorString
-import org.ossreviewtoolkit.downloader.VersionControlSystem
-import org.ossreviewtoolkit.model.Identifier
-import org.ossreviewtoolkit.model.Package
-import org.ossreviewtoolkit.model.PackageReference
-import org.ossreviewtoolkit.model.Project
+import org.ossreviewtoolkit.analyzer.PackageManagerResult
 import org.ossreviewtoolkit.model.ProjectAnalyzerResult
-import org.ossreviewtoolkit.model.RemoteArtifact
-import org.ossreviewtoolkit.model.Scope
-import org.ossreviewtoolkit.model.VcsInfo
-import org.ossreviewtoolkit.model.VcsType
-import org.ossreviewtoolkit.model.collectDependencies
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.RepositoryConfiguration
+import org.ossreviewtoolkit.model.utils.DependencyGraphBuilder
 import org.ossreviewtoolkit.utils.common.CommandLineTool
 import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.common.stashDirectories
@@ -71,6 +61,8 @@ class Bower(
         ) = Bower(type, analysisRoot, analyzerConfig, repoConfig)
     }
 
+    private val graphBuilder = DependencyGraphBuilder(BowerDependencyHandler())
+
     override fun beforeResolution(definitionFiles: List<File>) = BowerCommand.checkVersion()
 
     override fun resolveDependencies(definitionFile: File, labels: Map<String, String>): List<ProjectAnalyzerResult> {
@@ -82,23 +74,26 @@ class Bower(
                 .getTransitiveDependencies()
                 .associateBy { checkNotNull(it.pkgMeta.name) }
 
-            val scopes = SCOPE_NAMES.mapTo(mutableSetOf()) { scopeName ->
-                Scope(
-                    name = scopeName,
-                    dependencies = projectPackageInfo.toPackageReferences(scopeName, packageInfoForName)
-                )
+            SCOPE_NAMES.forEach { scopeName ->
+                val dependencies = projectPackageInfo.getScopeDependencies(scopeName).map { dependencyName ->
+                    // Bower leaves out a dependency entry for a child if there exists a similar entry to its parent
+                    // entry with the exact same name and resolved target. This makes it necessary to retrieve the
+                    // information about the subtree rooted at the parent from that other entry containing the full
+                    // dependency information.
+                    // See https://github.com/bower/bower/blob/6bc778d/lib/core/Manager.js#L557 and below.
+                    projectPackageInfo.dependencies[dependencyName] ?: packageInfoForName.getValue(dependencyName)
+                }
+
+                graphBuilder.addDependencies(projectPackageInfo.toIdentifier(), scopeName, dependencies)
             }
 
-            val project = projectPackageInfo.toProject(definitionFile, scopes)
-
-            val referencedPackages = scopes.collectDependencies()
-            val packages = packageInfoForName.values
-                .map { it.toPackage() }
-                .filterTo(mutableSetOf()) { it.id in referencedPackages }
-
-            return listOf(ProjectAnalyzerResult(project, packages))
+            val project = projectPackageInfo.toProject(definitionFile, SCOPE_NAMES)
+            return listOf(ProjectAnalyzerResult(project, emptySet()))
         }
     }
+
+    override fun createPackageManagerResult(projectResults: Map<File, List<ProjectAnalyzerResult>>) =
+        PackageManagerResult(projectResults, graphBuilder.build(), graphBuilder.packages())
 
     private fun getProjectPackageInfo(workingDir: File): PackageInfo {
         BowerCommand.run(workingDir, "--allow-root", "install").requireSuccess()
@@ -130,66 +125,3 @@ private fun PackageInfo.getTransitiveDependencies(): List<PackageInfo> {
 
     return result
 }
-
-private fun PackageInfo.toId() =
-    Identifier(
-        type = "Bower",
-        namespace = "",
-        name = pkgMeta.name.orEmpty(),
-        version = pkgMeta.version.orEmpty()
-    )
-
-private fun PackageInfo.toVcsInfo() =
-    VcsInfo(
-        type = VcsType.forName(pkgMeta.repository?.type.orEmpty()),
-        url = pkgMeta.repository?.url ?: pkgMeta.source.orEmpty(),
-        revision = pkgMeta.resolution?.commit ?: pkgMeta.resolution?.tag.orEmpty()
-    )
-
-private fun PackageInfo.toPackage() =
-    Package(
-        id = toId(),
-        // See https://github.com/bower/spec/blob/master/json.md#authors.
-        authors = pkgMeta.authors.flatMap { parseAuthorString(it.name) }.mapNotNullTo(mutableSetOf()) { it.name },
-        declaredLicenses = setOfNotNull(pkgMeta.license?.takeUnless { it.isEmpty() }),
-        description = pkgMeta.description.orEmpty(),
-        homepageUrl = pkgMeta.homepage.orEmpty(),
-        binaryArtifact = RemoteArtifact.EMPTY,
-        sourceArtifact = RemoteArtifact.EMPTY, // TODO: implement me!
-        vcs = toVcsInfo()
-    )
-
-private fun PackageInfo.toProject(definitionFile: File, scopes: Set<Scope>) =
-    with(toPackage()) {
-        Project(
-            id = id,
-            definitionFilePath = VersionControlSystem.getPathInfo(definitionFile).path,
-            authors = authors,
-            declaredLicenses = declaredLicenses,
-            vcs = vcs,
-            vcsProcessed = processProjectVcs(definitionFile.parentFile, vcs, homepageUrl),
-            homepageUrl = homepageUrl,
-            scopeDependencies = scopes
-        )
-    }
-
-private fun PackageInfo.toPackageReferences(
-    scopeName: String,
-    packageInfoForName: Map<String, PackageInfo>
-): Set<PackageReference> =
-    getScopeDependencies(scopeName).mapTo(mutableSetOf()) { name ->
-        // Bower leaves out a dependency entry for a child if there exists a similar entry to its parent entry
-        // with the exact same name and resolved target. This makes it necessary to retrieve the information
-        // about the subtree rooted at the parent from that other entry containing the full dependency
-        // information.
-        // See https://github.com/bower/bower/blob/6bc778d/lib/core/Manager.js#L557 and below.
-        val childInfo = dependencies[name] ?: packageInfoForName.getValue(name)
-
-        PackageReference(
-            id = childInfo.toId(),
-            dependencies = childInfo.toPackageReferences(
-                scopeName = SCOPE_NAME_DEPENDENCIES,
-                packageInfoForName = packageInfoForName
-            )
-        )
-    }
